rename error_helpers -> startup_helpers
