+++ lib/src/startup_helpers.rs
